+++ com.unity.template-hd/Assets/Scripts/SimpleCameraController.cs
@@ -1,10 +1,4 @@
-﻿#if ENABLE_INPUT_SYSTEM && ENABLE_INPUT_SYSTEM_PACKAGE
-#define USE_INPUT_SYSTEM
-    using UnityEngine.InputSystem;
-    using UnityEngine.InputSystem.Controls;
-#endif
-
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace UnityTemplateProjects
 {
@@ -114,11 +108,8 @@ namespace UnityTemplateProjects
         
         void Update()
         {
-            Vector3 translation = Vector3.zero;
-
-#if ENABLE_LEGACY_INPUT_MANAGER
-
             // Exit Sample  
+
             if (Input.GetKey(KeyCode.Escape))
             {
                 Application.Quit();
@@ -151,21 +142,17 @@ namespace UnityTemplateProjects
             }
             
             // Translation
-            translation = GetInputTranslationDirection() * Time.deltaTime;
+            var translation = GetInputTranslationDirection() * Time.deltaTime;
 
             // Speed up movement when shift key held
             if (Input.GetKey(KeyCode.LeftShift))
             {
                 translation *= 10.0f;
             }
-
+            
             // Modify movement by a boost factor (defined in Inspector and modified in play mode through the mouse scroll wheel)
             boost += Input.mouseScrollDelta.y * 0.2f;
             translation *= Mathf.Pow(2.0f, boost);
-
-#elif USE_INPUT_SYSTEM 
-            // TODO: make the new input system work
-#endif
 
             m_TargetCameraState.Translate(translation);